FIX remove float hack from material index

+++ shader/test6/_uniform_material_indeces.glsl
@@ -1,3 +1,3 @@
 layout (std430, binding = SSBO_MATERIAL_INDECES) readonly buffer MaterialIndeces {
-  float u_materialIndeces[];
+  int u_materialIndeces[];
 };
