+++ src/asset/MaterialIndex.h
@@ -4,6 +4,6 @@
 struct MaterialIndex {
     // NOTE KI uint DOES NOT work well in vertex attrs; data gets corrupted
     // => use float
-    float m_materialIndex;
+    GLint m_materialIndex;
 };
 #pragma pack(pop)
